templates/python/requirements.txt
# Runtime dependencies — keep this file up to date with `pip freeze > requirements.txt`
# after installing packages into your .venv.

anthropic>=0.40.0
python-dotenv>=1.0.0
